requirements.txt
# E2E Testing Framework Adoption and Migration Analysis

# Core dependencies
python-dotenv==0.19.2
SQLAlchemy==1.4.41

# Data collection and analysis
PyDriller==2.6
GitPython==3.1.43
requests==2.31.0

# Data processing
pandas==2.2.3
numpy==2.2.1
openpyxl==3.1.5
python-dateutil==2.9.0.post0
pytz==2024.2
tzdata==2024.2
six==1.17.0

# Visualization
matplotlib==3.9.2
plotly==5.13.1
kaleido==0.2.1
networkx==3.1

# Web UI (manual labeling)
Flask==3.1.0
Jinja2==3.1.5
Werkzeug==3.1.3
click==8.1.8
blinker==1.9.0
itsdangerous==2.2.0
MarkupSafe==3.0.2
